test: demonstrate incorrect -s option on mcron program
The option is supposed to be able to take an optional argument, but if the
argument is not supplied (should default to 8) then the test, rather than
failing, is skipped with a friendly message in the log file.  The proper fix
will come with an upstream patch to GNU Guile, and a future version of Mcron.

* tests/schedule-2.sh: new test, new file
* Makefile.am: make sure to run the new test file