documentation: extensive editing of info manual after a note from Paul Vixie.
Paul has been in touch to say that the crontab format was not his invention,
and that his program was based off of V7's cron's functionality, not later
AT&T and Berkeley ones.

Thus the mcron manual is edited extensively to emphasise more the POSIX
standard crontab format, and to point out a more accurate history of cron
developments.

* doc/mcron.texi: small changes throughout the document.