Remove (next-week) from texinfo documentation, discuss equivalent as example.
It has been decided not to implement the (next-week) procedure in the program,
so we need to remove the references to it in the manual.

The reason for the omission is discussed in a new extended example section,
and code is provided to illustrate how the functionality can be implemented if
needed.

The change also makes some alteration to the flow of sections in the document,
and implicitly fixes some problems in the flow that the previous document had,
with regards to linking the GNU Free Documentation License in the document
structure.

   * doc/mcron.texi:  modifications as described.